requirements.txt
flask==2.2.2
tensorflow==2.11.0
pillow==9.3.0

